feat: refactor equations handling to use long double for improved precision

+++ include/LibraryInterface.h
@@ -11,11 +11,12 @@
 // Required function signatures to be implemented by the user
 extern "C" {
 // Function to evaluate the i-th equation of the system
-FUNCTION_EXPORT double evaluateFunction(int i, int n, const double *x);
+FUNCTION_EXPORT long double evaluateFunction(int i, int n,
+                                             const long double *x);
 
 // Function to evaluate the derivatives of the i-th equation
-FUNCTION_EXPORT void evaluateDerivatives(int i, int n, const double *x,
-                                         double *dfatx);
+FUNCTION_EXPORT void evaluateDerivatives(int i, int n, const long double *x,
+                                         long double *dfatx);
 
 // Get a human-readable name for this system (optional)
 FUNCTION_EXPORT const char *getName();
@@ -26,4 +27,4 @@ FUNCTION_EXPORT int getNumberOfEquations();
 // // Get recommended initial values (optional)
 // FUNCTION_EXPORT void getRecommendedInitialGuess(double *initialGuess);
 }
-#endif // __LIBRARYINTERFACE_H__
+#endif  // __LIBRARYINTERFACE_H__
